+++ src/components/leaflet/LeafletMap.jsx
@@ -2,7 +2,7 @@ import { MapContainer, TileLayer, Marker } from 'react-leaflet';
 import style from './leaflet.module.scss';
 import { Icon } from 'leaflet';
 import MarkerIcon from '../../assets/images/icon-location.svg';
-import { MyComponent } from '../../helpers/myComponent';
+import { MovingMap } from '../../helpers/movingMap';
 
 const customIcon = new Icon({
   iconUrl: MarkerIcon,
@@ -14,13 +14,12 @@ const customIcon = new Icon({
 const LeafletMap = (props) => {
   return (
     <MapContainer
-      // key={`${props.coordinates[0]}`}
       className={style.mapCont}
       center={props.coordinates}
       zoom={13}
       scrollWheelZoom={false}
     >
-      <MyComponent geoInfo={props.coordinates} />
+      <MovingMap geoInfo={props.coordinates} />
       <TileLayer
         attribution='&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors'
         url="https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png"
